Checkout this comment by me for ui reference:- [Feature Request]: Create a Admin dashboard page for content and report management. #13 (comment) - done similarly .
First of all git pull all the changes. - it should be up to date .

For UI reference you can checkout pocketbase.io. - checked it out and tried to do it.

And try breaking down the UI into components it will be easy to understand and maintainable in future. - did as best as i could i think the rest would be better for you to do.

Fix bugs - fixed bugs related to my code.

Folder structure - i tried to do it as you asked but i think you should rearange it as y prefer in the end.

replaced usereffect and used prisma except for logs had issues so i left it as is....

removing files .env.local, yarn.lock,package,bun - done